Revert "Seed Libc rng from system entropy (#43606)"
This reverts commit 5819775e5aa78d808132b24a7b15b02d11bbeefc.
